requirements.txt
djangorestframework>=3.10
